Bundled hero-assign
