+++ Anchor.toml
@@ -5,7 +5,6 @@ resolution = true
 skip-lint = false
 
 [programs.devnet]
-assign_hero = "7gBLDn72Cog7dBvN1LWfo6W36Q7vxcv7CqYAeHwfo3Y"
 assign_settlement = "42g6wojVK214btG2oUHg8vziW8UaUiQfPZ6K9kMGTCp2"
 build = "fkiWK1Wn6ouGcHb3icX4XGKynef5MpsTQ478ZMdgB1g"
 claim_loot = "4CjxHvNUpoCYomULBFTvmkTQPaNd9QDHPhZQ6eB9bZEf"
@@ -26,7 +25,6 @@ wait = "9F6qiZPUWN3bCnr5uVBwSmEDf8QcAFHNSVDH8L7AkZe4"
 worker = "BExuAEwcKxKeqHSN8C1WetUAd6Tm71cZEiP8EBSrH55T"
 
 [programs.localnet]
-assign_hero = "7gBLDn72Cog7dBvN1LWfo6W36Q7vxcv7CqYAeHwfo3Y"
 assign_settlement = "42g6wojVK214btG2oUHg8vziW8UaUiQfPZ6K9kMGTCp2"
 build = "fkiWK1Wn6ouGcHb3icX4XGKynef5MpsTQ478ZMdgB1g"
 bump_location_allocator = "C2H1sb7ZVpgEZFWqXujRK3rx5C2543GNN251wmgfbhUH"
